+++ Sources/Ether/FixInstall.swift
@@ -36,12 +36,12 @@ public final class FixInstall: Command {
         let fixing = context.console.loadingBar(title: "Fixing Instillation")
         _ = fixing.start(on: context.container)
         
-        _ = try Process.execute("rm", ["--rf", ".build"])
+        _ = try Process.execute("rm", ["-rf", ".build"])
         _ = try Process.execute("swift", ["package", "update"])
         _ = try Process.execute("swift", ["package", "resolve"])
         
         fixing.succeed()
         
-        return context.container.eventLoop.newSucceededFuture(result: ())
+        return context.container.future()
     }
 }